All unit tests are functioning as inteded.
