--- internal/controller/test/logging_test_go
@@ -1,22 +1,0 @@
-package controller_test
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-var _ = Describe("BitwardenSecretReconciler", func() {
-	Describe("LogWarning", func() {
-		It("should log a warning correctly", func() {})
-	})
-
-	Describe("LogError", func() {
-		It("should log and update status", func() {})
-		It("should handle status update failure", func() {})
-	})
-
-	Describe("LogCompletion", func() {
-		It("should log and update status", func() {})
-		It("should handle status update failure", func() {})
-	})
-})
