+++ internal/controller/test/reconciler_success_test.go
@@ -1,0 +1,142 @@
+package controller_test
+
+import (
+	"time"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+
+	corev1 "k8s.io/api/core/v1"
+
+	apimeta "k8s.io/apimachinery/pkg/api/meta"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	"k8s.io/apimachinery/pkg/types"
+
+	sdk "github.com/bitwarden/sdk-go"
+	operatorsv1 "github.com/bitwarden/sm-kubernetes/api/v1"
+	"github.com/bitwarden/sm-kubernetes/internal/controller"
+	"github.com/bitwarden/sm-kubernetes/internal/controller/test/testutils"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+	//+kubebuilder:scaffold:imports
+)
+
+var _ = Describe("BitwardenSecret Reconciler - Success Tests", Ordered, func() {
+	var (
+		namespace string
+		fixture   testutils.TestFixture
+	)
+
+	BeforeEach(func() {
+		fixture = *testutils.NewTestFixture(testContext, envTestRunner)
+		namespace = fixture.CreateNamespace()
+	})
+
+	AfterAll(func() {
+		fixture.Cancel()
+	})
+
+	AfterEach(func() {
+		fixture.Teardown()
+	})
+
+	It("should complete a successful sync", func() {
+		fixture.SetupDefaultCtrlMocks(false, nil)
+
+		_, err := fixture.CreateDefaultAuthSecret(namespace)
+		Expect(err).NotTo(HaveOccurred())
+
+		bwSecret, err := fixture.CreateDefaultBitwardenSecret(namespace, fixture.SecretMap)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(bwSecret).NotTo(BeNil())
+
+		req := reconcile.Request{NamespacedName: types.NamespacedName{Name: testutils.BitwardenSecretName, Namespace: namespace}}
+
+		result, err := fixture.Reconciler.Reconcile(fixture.Ctx, req)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(result.RequeueAfter).To(Equal(time.Duration(fixture.Reconciler.RefreshIntervalSeconds) * time.Second))
+
+		Eventually(func(g Gomega) {
+			// Verify created secret
+			createdTargetSecret := &corev1.Secret{}
+			g.Expect(fixture.K8sClient.Get(fixture.Ctx, types.NamespacedName{Name: testutils.SynchronizedSecretName, Namespace: namespace}, createdTargetSecret)).Should(Succeed())
+			g.Expect(createdTargetSecret.Labels[controller.LabelBwSecret]).To(Equal(string(bwSecret.UID)))
+			g.Expect(createdTargetSecret.Type).To(Equal(corev1.SecretTypeOpaque))
+			g.Expect(len(createdTargetSecret.Data)).To(Equal(testutils.ExpectedNumOfSecrets))
+
+			// Verify annotations
+			g.Expect(createdTargetSecret.Annotations[controller.AnnotationSyncTime]).NotTo(BeEmpty())
+			g.Expect(createdTargetSecret.Annotations[controller.AnnotationCustomMap]).NotTo(BeEmpty())
+
+			// Verify SuccessfulSync condition and LastSuccessfulSyncTime
+			updatedBwSecret := &operatorsv1.BitwardenSecret{}
+			g.Expect(fixture.K8sClient.Get(fixture.Ctx, types.NamespacedName{Name: testutils.BitwardenSecretName, Namespace: namespace}, updatedBwSecret)).Should(Succeed())
+			condition := apimeta.FindStatusCondition(updatedBwSecret.Status.Conditions, "SuccessfulSync")
+			g.Expect(condition).NotTo(BeNil())
+			g.Expect(condition.Status).To(Equal(metav1.ConditionTrue))
+			g.Expect(updatedBwSecret.Status.LastSuccessfulSyncTime.Time).NotTo(BeZero())
+		}).Should(Succeed())
+	})
+
+	It("should skip reconciliation when last sync is within refresh interval", func() {
+		fixture.SetupDefaultCtrlMocks(false, nil)
+
+		_, err := fixture.CreateDefaultAuthSecret(namespace)
+		Expect(err).NotTo(HaveOccurred())
+
+		bwSecret, err := fixture.CreateDefaultBitwardenSecret(namespace, fixture.SecretMap)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(bwSecret).NotTo(BeNil())
+
+		// Update status with LastSuccessfulSyncTime, retrying on conflicts
+		syncTime := time.Now().UTC()
+		Eventually(func(g Gomega) {
+			// Fetch the latest version of bwSecret
+			latestBwSecret := &operatorsv1.BitwardenSecret{}
+			g.Expect(fixture.K8sClient.Get(fixture.Ctx, types.NamespacedName{Name: testutils.BitwardenSecretName, Namespace: namespace}, latestBwSecret)).Should(Succeed())
+
+			// Update status
+			latestBwSecret.Status = operatorsv1.BitwardenSecretStatus{
+				LastSuccessfulSyncTime: metav1.Time{Time: syncTime},
+			}
+			g.Expect(fixture.K8sClient.Status().Update(fixture.Ctx, latestBwSecret)).Should(Succeed())
+		}).WithTimeout(10 * time.Second).WithPolling(100 * time.Millisecond).Should(Succeed())
+
+		req := reconcile.Request{NamespacedName: types.NamespacedName{Name: testutils.BitwardenSecretName, Namespace: namespace}}
+
+		result, err := fixture.Reconciler.Reconcile(fixture.Ctx, req)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(result).To(Equal(reconcile.Result{}))
+	})
+
+	It("should skip sync when no changes from Bitwarden API", func() {
+		// Override mocks to return no changes
+		noChangesResponse := sdk.SecretsSyncResponse{
+			HasChanges: false,
+			Secrets:    []sdk.SecretResponse{},
+		}
+
+		fixture.SetupDefaultCtrlMocks(false, &noChangesResponse)
+
+		_, err := fixture.CreateDefaultAuthSecret(namespace)
+		Expect(err).NotTo(HaveOccurred())
+
+		bwSecret, err := fixture.CreateDefaultBitwardenSecret(namespace, fixture.SecretMap)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(bwSecret).NotTo(BeNil())
+
+		req := reconcile.Request{NamespacedName: types.NamespacedName{Name: testutils.BitwardenSecretName, Namespace: namespace}}
+
+		result, err := fixture.Reconciler.Reconcile(fixture.Ctx, req)
+		Expect(err).NotTo(HaveOccurred())
+		Expect(result.RequeueAfter).To(Equal(time.Duration(fixture.Reconciler.RefreshIntervalSeconds) * time.Second))
+
+		Eventually(func(g Gomega) {
+			// Verify no SuccessfulSync condition (no sync occurred)
+			createdSecret := &operatorsv1.BitwardenSecret{}
+			g.Expect(fixture.K8sClient.Get(fixture.Ctx, types.NamespacedName{Name: testutils.BitwardenSecretName, Namespace: namespace}, createdSecret)).Should(Succeed())
+			condition := apimeta.FindStatusCondition(createdSecret.Status.Conditions, "SuccessfulSync")
+			g.Expect(condition).To(BeNil())
+		})
+	})
+})
