--- internal/controller/test/secret_manager_test.go
@@ -1,17 +1,0 @@
-package controller_test
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-)
-
-var _ = Describe("BitwardenSecretReconciler", func() {
-	Describe("PullSecretManagerSecretDeltas", func() {
-		It("should handle client creation failure", func() {})
-		It("should handle authentication failure", func() {})
-		It("should handle sync failure", func() {})
-		It("should handle a nil response", func() {})
-		It("should handle no changes", func() {})
-		It("should handle changes", func() {})
-		It("should ensure client closure", func() {})
-	})
-})
