--- internal/controller/test/secret_utils_test.go
@@ -1,33 +1,0 @@
-package controller_test
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-)
-
-var _ = Describe("BitwardenSecretReconciler", func() {
-	Describe("CreateK8sSecret", func() {
-		It("should create a basic secret", func() {})
-		It("should handle an empty BitwardenSecret", func() {})
-	})
-
-	Describe("ApplySecretMap", func() {
-		It("should apply secrets with no mapping and OnlyMappedSecrets=false", func() {})
-		It("should apply no secrets with no mapping and OnlyMappedSecrets=true", func() {})
-		It("should apply secrets with partial mapping and OnlyMappedSecrets=false", func() {})
-		It("should apply only mapped secrets with partial mapping and OnlyMappedSecrets=true", func() {})
-		It("should handle a nil Data field", func() {})
-	})
-
-	Describe("FindSecretMapByBwSecretId", func() {
-		It("should return a matching SecretMap", func() {})
-		It("should return empty for a non-matching ID", func() {})
-		It("should return empty for a nil SecretMap", func() {})
-	})
-
-	Describe("SetK8sSecretAnnotations", func() {
-		It("should set annotations with no SecretMap", func() {})
-		It("should set annotations with a SecretMap", func() {})
-		It("should handle JSON marshal failure", func() {})
-		It("should preserve existing annotations", func() {})
-	})
-})
