--- internal/controller/test/setup_test.go
@@ -1,12 +1,0 @@
-package controller_test
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-)
-
-var _ = Describe("BitwardenSecretReconciler", func() {
-	Describe("SetupWithManager", func() {
-		It("should configure the controller correctly", func() {})
-		It("should handle manager failure", func() {})
-	})
-})
